+++ packages/patrol/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - **Breaking:** Change signature of `PatrolTester.pumpAndSettle()` method to use
   named arguments (#657)
+- Fix `PatrolTester.dragUntilVisible()` not calling `first` on its `Finder view`
+  parameter (#656)
 
 ## 0.7.6